+++ src/main/java/ghstats/api/achievements/SqlAchievementsRepository.java
@@ -13,7 +13,7 @@ public class SqlAchievementsRepository implements AchievementsRepository {
     }
 
     @Override
-    public Mono<Integer> saveAchievement(String achievementId, AchievementUnlocked achievementUnlocked) {
+    public Mono<Long> saveAchievement(String achievementId, AchievementUnlocked achievementUnlocked) {
         return databaseClient.sql("INSERT IGNORE INTO `achievements_unlocked` (`user`, `commit_id`, `achievement_id`) VALUES (?, ?, ?)")
                 .bind(0, achievementUnlocked.commit().author().userName().value())
                 .bind(1, achievementUnlocked.commit().id().value())
